Fixed Numexpr string test to work regardless of machine word size.
git-svn-id: http://www.pytables.org/svn/pytables/branches/professional@2170 1b98710c-d8ec-0310-ae81-f5f2bcd8cb94

+++ tables/numexpr/tests/test_numexpr.py
@@ -384,7 +384,7 @@ class test_strings(NumpyTestCase):
 
     def check_add_numeric_array(self):
         sarr = self.str_array1
-        narr = arange(len(sarr))
+        narr = arange(len(sarr), dtype='int32')
         expr = 'sarr >= narr'
         self.assert_missing_op('ge_bsi', expr, locals())
